Small updates for more language consistency
Although the translation was good enough, I wanted to improve slightly on the consistency of the language that is being used in the Xperia Z settings panel. Also I want to suggest to rename the Xperia Z Settings menu to Device settings for a more consistent feel. Although that may be a personal preference.

If you don't care for that last change you can also change the "Apparaatinstellingen" in "Xperia Z instellingen" in the Dutch translation. But of course, that is up to you :).